Refactor weather app views and widgets
This commit refactors the weather app views and widgets by renaming and reorganizing some files. Specifically, the `WeatherDisplayPage` view was renamed to `WeatherBody`, and a new `HomeView` was added to serve as the app's home page. Additionally, the `NoWeatherBody` widget was moved to a new `widgets` directory, and a new `WeatherInfoBody` widget was added to display weather information.